+++ crates/hooks/src/lib.rs
@@ -5,6 +5,7 @@ mod animation;
 mod rope_editor;
 mod shader_uniforms;
 mod text_editor;
+mod theming;
 
 mod use_accessibility;
 mod use_animation;
@@ -23,6 +24,7 @@ pub use animation::*;
 pub use rope_editor::*;
 pub use shader_uniforms::*;
 pub use text_editor::*;
+pub use theming::*;
 
 pub use use_accessibility::*;
 pub use use_animation::*;
